[Fix]Home / [Feature - components] UserMain, OwnerMain
