+++ src/main/frontend/src/pages/StoreInto.jsx
@@ -4,6 +4,7 @@ import Footer from "../components/Footer"
 import PropTypes from 'prop-types';
 import { Tab, Tabs, Box, Typography, Stack, Grid, InputBase} from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
+import { Link } from 'react-router-dom'
 
 
 function CustomTabPanel(props) {
@@ -50,6 +51,27 @@ export default function Store() {
   
   return (
     <Box sx={{ margin: -1 }}>
+
+<Grid container>
+          <Grid item xs/>
+          <Grid container sx={{ position: 'relative', border: 1, borderColor: 'rgba(255, 0, 0, 0)', justifyContent: 'center', alignItems: 'center' }}>
+              <Grid className="centerBody" container columnSpacing={{ xs: 2, sm: 2}} sx={gridStyle}>
+              <Box sx={{...boxStyle, position: 'relative', width: { xs: '90%', sm: '47%' }, height: '120px', marginX: 'auto'}}>
+                <div>
+                  <img src={'/img/01.jpg'} style={{ width: '20%', height: '100%', position: 'absolute', top: 0, left: 0 }} />
+                  <ul style={{ position: 'absolute', top: '50%', left: '40%', transform: 'translate(-50%, -50%)', padding: 0, margin: 0 }}>
+                    <li style={{ listStyleType: 'none' }}>가게 주소</li>
+                    <li style={{ listStyleType: 'none' }}>찜순</li>
+                    <li style={{ listStyleType: 'none' }}>최소 주문 금액</li>
+                    <li style={{ listStyleType: 'none' }}>평점순</li>
+                    <li style={{ listStyleType: 'none' }}>리뷰순</li>
+                  </ul>
+                </div> 
+                </Box>
+              </Grid>
+            </Grid>
+          <Grid item xs />
+        </Grid>
       
       <Grid container></Grid>
       <Box sx={{ borderBottom: 1, borderColor: 'divider',  display: 'flex', justifyContent: 'center' }}>
@@ -72,27 +94,7 @@ export default function Store() {
         </Grid>
       </Grid>
       <CustomTabPanel value={value} index={0}>
-        Item 정보
-        <Grid container>
-          <Grid item xs/>
-          <Grid container sx={{ position: 'relative', border: 1, borderColor: 'rgba(255, 0, 0, 0)', justifyContent: 'center', alignItems: 'center' }}>
-              <Grid className="centerBody" container columnSpacing={{ xs: 2, sm: 2}} sx={gridStyle}>
-                <Box sx={{...boxStyle, position: 'relative', width: { xs: '90%', sm: '47%' }, height: '120px', marginX: 'auto'}}>
-                <div>
-                  <img src={'/img/01.jpg'} style={{ width: '20%', height: '100%', position: 'absolute', top: 0, left: 0 }} />
-                  <ul style={{ position: 'absolute', top: '50%', left: '40%', transform: 'translate(-50%, -50%)', padding: 0, margin: 0 }}>
-                    <li style={{ listStyleType: 'none' }}>가게 주소</li>
-                    <li style={{ listStyleType: 'none' }}>찜순</li>
-                    <li style={{ listStyleType: 'none' }}>최소 주문 금액</li>
-                    <li style={{ listStyleType: 'none' }}>평점순</li>
-                    <li style={{ listStyleType: 'none' }}>리뷰순</li>
-                  </ul>
-                </div> 
-                </Box>
-              </Grid>
-            </Grid>
-          <Grid item xs />
-        </Grid>
+        
       </CustomTabPanel>
       <CustomTabPanel value={value} index={1}>
         Item 메뉴
