+++ src/main/frontend/src/pages/StoreRegister.jsx
@@ -7,7 +7,7 @@ import { Link, useNavigate } from 'react-router-dom';
 import { findPostcode } from '../utils/AddressUtil'; 
 import { getCurrentUser, register } from '../utils/firebase';
 import { extractDataFromFormData } from '../utils/userInfo';
-import { useMutation } from '@tanstack/react-query';
+
 import axios from 'axios';
 
 const defaultTheme = createTheme();
